Fix randomly failling test
When the dependency has not rubygem, meaning it is unresolved, there is
no way to guarantee the order. So we cannot rely on array look on tests.

This is probably falling only a few times, due to the records all get
inserted inside a transaction, which probably wont have fixed order.